# api/openapi.yaml
openapi: 3.0.3
info:
  title: GophKeeper API
  version: 0.1.0
  description: HTTP API for the GophKeeper password manager project.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/auth/refresh:
    post:
      summary: Rotate a refresh token and return a new session
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Session refreshed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    AuthUnavailable:
      description: Auth service is not configured on this server.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    KDFParams:
      type: object
      required: [algorithm, memory_kib, iterations, parallelism, key_length]
      properties:
        algorithm:
          type: string
          enum: [argon2id]
        memory_kib:
          type: integer
          format: int32
        iterations:
          type: integer
          format: int32
        parallelism:
          type: integer
          format: int32
        key_length:
          type: integer
          format: int32
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    AuthSessionResponse:
      type: object
      required: [user_id, login, access_token, access_expires_at, refresh_token, refresh_expires_at]
      properties:
        user_id:
          type: string
        login:
          type: string
        access_token:
          type: string
        access_expires_at:
          type: string
          format: date-time
        refresh_token:
          type: string
        refresh_expires_at:
          type: string
          format: date-time
        auth_salt:
          type: string
          format: byte
        vault_salt:
          type: string
          format: byte
        kdf_params:
          $ref: "#/components/schemas/KDFParams"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string